+++ src/mantle/facade/class-route.php
@@ -36,7 +36,7 @@ namespace Mantle\Facade;
  * @method static void rest_api(string $namespace, callable|string $callback, callable|array|string $args = [])
  * @method static void model(string $model, string $controller)
  * @method static void sync_routes_to_url_generator()
- * @method static \Mantle\Http\Routing\Router rename_route(string $old_name, string $new_name)
+ * @method static \static rename_route(string $old_name, string $new_name)
  * @method static bool has_group_stack()
  * @method static array get_group_stack()
  * @method static void group(array $attributes, \Closure|string $routes)